requirements.txt
google-generativeai
python-dotenv
grpcio==1.60.1
pillow